scripts/requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
pandas
click
